+++ bevy_pretty_nice_input_derive/src/lib.rs
@@ -10,11 +10,13 @@ pub fn derive_action(input: TokenStream) -> TokenStream {
     derive_action::derive_action(input)
 }
 
+/// Usage: `input!(action, Axis_D[bindings], [conditions])`
 #[proc_macro]
 pub fn input(input: TokenStream) -> TokenStream {
     input::input_impl(input)
 }
 
+/// Usage: `input_transition!(action: states <=/<=>/=> states, Axis_D[bindings], [conditions])`
 #[proc_macro]
 pub fn input_transition(input: TokenStream) -> TokenStream {
     input_transition::input_transition_impl(input)