+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 0.2.0
 
+### Added
+
+- `ActionData::as_1d_ok`/`as_2d_ok`/`as_3d_ok` which return `Result<_, BevyError>`
+- New example for a physics-based (bevy_rapier3D) first-person player controller
+- Macro format doc comments
+
 ### Changed
 
 - Renamed the crate to use underscores.